update Cookie.token to be uint64_t instead of unsigned int this is in following with the standard set by NIST SP 800-63B

+++ example/main.cpp
@@ -187,7 +187,7 @@ int createSingleUser(){
 void cookieTest(){
     initCookieDB();
     cout << "Enter token value\n> ";
-    unsigned int token;
+    uint64_t token;
     cin >> token;
 
     login l;
